+++ package.json
@@ -9,24 +9,24 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^6.3.0",
-    "@fortawesome/free-brands-svg-icons": "^6.3.0",
-    "@fortawesome/free-solid-svg-icons": "^6.3.0",
-    "@fortawesome/react-fontawesome": "^0.2.0",
     "@next/font": "13.1.6",
     "@types/node": "18.14.0",
     "@types/react": "18.0.28",
     "@types/react-dom": "18.0.11",
     "eslint": "8.34.0",
     "eslint-config-next": "13.1.6",
-    "framer-motion": "^9.0.4",
     "next": "13.1.6",
     "react": "18.2.0",
-    "react-custom-scrollbars": "^4.2.1",
     "react-dom": "18.2.0",
+    "typescript": "4.9.5",
+    "@fortawesome/fontawesome-svg-core": "^6.3.0",
+    "@fortawesome/free-brands-svg-icons": "^6.3.0",
+    "@fortawesome/free-solid-svg-icons": "^6.3.0",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "framer-motion": "^9.0.4",
+    "react-custom-scrollbars": "^4.2.1",
     "react-is": "^18.2.0",
-    "styled-components": "^5.3.6",
-    "typescript": "4.9.5"
+    "styled-components": "^5.3.6"
   },
   "devDependencies": {
     "@types/react-custom-scrollbars": "^4.0.10",
