+++ src/pages/about.tsx
@@ -1,5 +1,3 @@
-"use client"
-
 import Logo from "../components/Logo";
 import NavBar from "../components/NavBar";
 import AboutText from "../components/AboutText";
