+++ src/pages/index.tsx
@@ -1,4 +1,3 @@
-"use client"
 import Logo from "../components/Logo";
 import NavBar from "../components/NavBar";
 import Welcome from "../components/Welcome";
