+++ src/pages/projects.tsx
@@ -1,5 +1,3 @@
-"use client"
-
 import NavBar from "../components/NavBar";
 import { ProjectContainer, ProjectItem } from "../components/Project";
 import { Scrollbars } from "react-custom-scrollbars";
