+++ src/styles/globals.scss
@@ -6,8 +6,6 @@
   padding: 0;
   margin: 0;
   box-sizing: border-box;
-
-  font-family: 'Josefin Sans', sans-serif;
 }
 
 body {
